client code refactoring for generatorheader
