+++ client/src/components/GeneratorApp/GeneratorApp.js
@@ -16,7 +16,7 @@ import React from 'react';
 
 import CodeStructure from './CodeStructure';
 import CodeViewer from './CodeViewer';
-import Header from './Header/Header';
+import GeneratorAppHeader from './Header/GeneratorAppHeader';
 
 import "../../styles_CSS/Generator/GeneratorApp.css";
 
@@ -95,7 +95,7 @@ class GeneratorApp extends React.Component {
       return (
         <div>
           {/* Renders the loader if loading is true. This means that the template is not back yet */}
-          <Header
+          <GeneratorAppHeader
             newProjectNameHandlerFromParent={this.newProjectNameHandler}
             sentInCurrentProjectName={this.state.currentProjectName}
             sentInDownloadable={this.state.downloadable}
@@ -109,7 +109,7 @@ class GeneratorApp extends React.Component {
     return (
       <div>
         {/* Renders the regular view of the file and code structure */}
-        <Header
+        <GeneratorAppHeader
           newProjectNameHandlerFromParent={this.newProjectNameHandler}
           sentInCurrentProjectName={this.state.currentProjectName}
           sentInDownloadable={this.state.downloadable}
